Doesn't try to delete root dir by accident

+++ src/hermes/acquisition/tpx3serval.py
@@ -55,9 +55,15 @@ def verify_working_dir(run_configs):
     status_files_dir = os.path.join(run_dir, run_configs.WorkingDir.path_to_status_files)      # Setting path for status files
     init_files_dir = os.path.join(run_dir, run_configs.WorkingDir.path_to_init_files)          # Setting path to save init files
 
+
     if os.path.exists(run_dir):
+
+        if run_dir in ["", "/"]:
+            raise RuntimeError("verify_working_dir() refusing to operate on root directory")
+            
         print(f"Duplicate directory exists. Cleaning existing directory: {run_dir}")
         shutil.rmtree(run_dir)
+        
 
     print(f"Verifying dir:{run_dir} and its sub-dirs")
     # List of directories to check and create if they don't exist
